+++ resources/assets/js/app.js
@@ -1,12 +1,16 @@
 import Vue          from 'vue'
 import VueRouter    from 'vue-router'
-import              '../vue/directives/PathValidator';
+import              '../vue/directives/validate-path';
+import              '../vue/directives/field-attrs';
 
 // state
 import admin        from './state/admin'
 import store        from './state/store'
 import config       from './functions/config'
 config();
+
+// functions
+import Helpers      from './functions/helpers';
 
 // components
 import Home         from '../vue/pages/Home.vue'
@@ -62,7 +66,7 @@ router.map(routes);
 router.afterEach(function (transition) {
 	if (transition.to.path.indexOf('/run') !== 0)
 	{
-		document.title = 'Sketchpad - ' + transition.to.matched[0].handler.component.name.toLowerCase();
+		Helpers.setTitle(transition.to.matched[0].handler.component.name);
 	}
 })